Split reference types
