+++ packages/types/src/index.ts
@@ -1,3 +1,4 @@
 export * from './discord.js'
+export * from './discord/reference.js'
 export * from './discordeno.js'
 export * from './shared.js'
